+++ Code/diorama_private_files.cmake
@@ -16,6 +16,7 @@ set(FILES
     Source/Clients/SkeletalClip.h
     Source/Clients/AsepriteImport.h
     Source/Clients/AsepriteImport.cpp
+    Source/Clients/AsepriteSheetData.h
     Source/Clients/AsepriteBinary.h
     Source/Clients/AsepriteBinary.cpp
     Source/Clients/Collision2D.h